change limit and offset data type to Long in DaoFilter

+++ thirdeye-persistence/src/main/java/ai/startree/thirdeye/datalayer/bao/TaskManagerImpl.java
@@ -246,7 +246,7 @@ public class TaskManagerImpl implements TaskManager {
     final long startTime = System.nanoTime();
     final List<TaskDTO> tasksToBeDeleted = filter(new DaoFilter()
         .setPredicate(Predicate.LT("createTime", formattedDate))
-        .setLimit(limit)
+        .setLimit((long) limit)
     );
 
     /* Delete each task */
